new branch based off of sd-webui
not quite yet working, but functionally the chat commands will work the same, but be piped through the webui, allowing for more complex features and updates as the webui separately updates

why use selenium and not just hook into websockets directly? That's a good question, I'd offer you a cookie, but I neither have cookies, nor a satisfying answer beyond my apparent inability to perform the magic required for messing around with the websockets such guis are built on.